+++ src/PRo3D.Viewer/Viewer-Model.fs
@@ -106,6 +106,7 @@ type ViewerAction =
     | NewScene
     | KeyDown                         of key : Aardvark.Application.Keys
     | KeyUp                           of key : Aardvark.Application.Keys      
+    | ResizeMainControl               of V2i * string
     | SetKind                         of TrafoKind
     | SetInteraction                  of Interactions        
     | SetMode                         of TrafoMode
@@ -131,7 +132,7 @@ type ViewerAction =
     | Logging                         of string * ViewerAction
     | ThreadsDone                     of string    
     | SnapshotThreadsDone             of string
-    | OnResize                        of V2i
+    | OnResize                        of V2i * string
     | StartDragging                   of V2i * MouseButtons
     | Dragging                        of V2i
     | EndDragging                     of V2i * MouseButtons
@@ -322,7 +323,7 @@ type Model = {
     picking          : bool
     ctrlFlag         : bool
     frustum          : Frustum
-    viewPortSize     : V2i
+    viewPortSize     : HashMap<string, V2i>
     overlayFrustum   : Option<Frustum>
     
     minervaModel     : PRo3D.Minerva.MinervaModel
@@ -520,7 +521,7 @@ module Viewer =
             //instrumentFrustum = Frustum.perspective 60.0 0.1 10000.0 1.0
             viewerMode = ViewerMode.Standard                
             footPrint = ViewPlanModel.initFootPrint
-            viewPortSize = V2i.One
+            viewPortSize = HashMap.empty
 
             arnoldSnapshotThreads = ThreadPool.empty
             showExplorationPoint = startupArgs.showExplorationPoint