+++ src/PRo3D.Viewer/Viewer/Viewer.fs
@@ -65,7 +65,7 @@ module UserFeedback =
             msg     = ViewerAction.NoAction ""
         }
 
-    let createWorker (feedback: UserFeedback<'a>) =
+    let createWorker (feedback: UserFeedback<ViewerAction>) =
         proclist {
             yield UpdateUserFeedback ""
             yield UpdateUserFeedback feedback.text
@@ -529,7 +529,7 @@ module ViewerApp =
             | Some selected ->                             
                 let f = (fun x ->
                     let a = x |> Leaf.toAnnotation
-                    let a = AnnotationProperties.update a msg 
+                    let a = AnnotationProperties.update a m.scene.referenceSystem.planet msg
 
                     //update true thickness computation on dip angle change
                     let a = 
@@ -550,7 +550,7 @@ module ViewerApp =
         | BookmarkMessage msg,_,_ ->  
             //Log.warn "[Viewer] bookmarks animation %A" m.navigation.camera.view.Location
 
-            let m', bm = Bookmarks.update m.scene.bookmarks msg _navigation m
+            let m', bm = Bookmarks.update m.scene.bookmarks m.scene.referenceSystem.planet msg _navigation m
             let animation = 
                 match msg with
                 | BookmarkAction.GroupsMessage k ->
@@ -838,7 +838,10 @@ module ViewerApp =
         | KeyDown k, _, _ ->
             let m =
                 match k with
-                | Aardvark.Application.Keys.LeftShift -> { m with shiftFlag = true}
+                | Aardvark.Application.Keys.LeftShift -> 
+                    let m = { m with shiftFlag = true}
+                    Log.line "[Viewer] ShiftFlag %A" m.shiftFlag
+                    m
                 | _ -> m
           
             let drawingAction =
@@ -1021,8 +1024,12 @@ module ViewerApp =
         | KeyUp k, _,_ ->               
             let m =
                 match k with
-                | Aardvark.Application.Keys.LeftShift -> { m with shiftFlag = false}
+                | Aardvark.Application.Keys.LeftShift -> 
+                    let m = { m with shiftFlag = false}
+                    Log.line "[Viewer] ShiftFlag %A" m.shiftFlag
+                    m                    
                 | _ -> m
+
             match k with
             | Aardvark.Application.Keys.LeftCtrl -> 
                 match m.interaction with
@@ -1068,7 +1075,7 @@ module ViewerApp =
             //Log.line "config message %A" a
             let c' = ConfigProperties.update m.scene.config a
             let m = Optic.set (Model.scene_ >-> Scene.config_) c' m
-            
+    
             match a with                   
             | ConfigProperties.Action.SetNearPlane _ | ConfigProperties.Action.SetFarPlane _ ->
                 let fov = m.frustum |> Frustum.horizontalFieldOfViewInDegrees
@@ -1076,7 +1083,7 @@ module ViewerApp =
                 let f' = Frustum.perspective fov c'.nearPlane.value c'.farPlane.value asp                    
 
                 { m with frustum = f' }
-                | _ -> m
+            | _ -> m
         | SetMode d,_,_ -> 
             { m with trafoMode = d }
         | SetKind d,_,_ -> 
@@ -1217,9 +1224,12 @@ module ViewerApp =
                 { m with overlayFrustum = None; linkingModel = linking' } //navigation = { m.navigation with camera = camera' }}
             | _ -> 
                 { m with linkingModel = PRo3D.Linking.LinkingApp.update m.linkingModel a }
-        | OnResize a,_,_ ->              
+        | OnResize (a,id),_,_ ->              
             Log.line "[RenderControl Resized] %A" a
-            { m with frustum = m.frustum |> Frustum.withAspect(float a.X / float a.Y) }
+            { m with frustum = m.frustum |> Frustum.withAspect(float a.X / float a.Y); viewPortSize = HashMap.add id a m.viewPortSize }
+        | ResizeMainControl(a,id),_,_ -> 
+            printfn "resize %A" (a,id)
+            { m with frustum = m.frustum |> Frustum.withAspect(float a.X / float a.Y); viewPortSize = HashMap.add id a m.viewPortSize }
         | SetTextureFiltering b,_,_ ->
             {m with filterTexture = b}
        // | TestHaltonRayCasting _,_,_->
@@ -1345,7 +1355,7 @@ module ViewerApp =
         |> Sg.noEvents
         |> Sg.trafo(trafo)
     
-    let renderControlAttributes (m: AdaptiveModel) = 
+    let renderControlAttributes (id : string) (m: AdaptiveModel) = 
         let renderControlAtts (model: AdaptiveNavigationModel) =
             amap {
                 let! state = model.navigationMode
@@ -1370,7 +1380,17 @@ module ViewerApp =
                 onKeyDown (KeyDown)
                 onKeyUp   (KeyUp)        
                 clazz "mainrendercontrol"
-              //  onResize  (OnResize)
+                onEvent "resizeControl"  [] (
+                    fun p -> 
+                        match p with
+                        | w::h::[] -> 
+                            let w : int = Pickler.json.UnPickleOfString w
+                            let h : int = Pickler.json.UnPickleOfString h
+                            printfn "%A" (w,h)
+                            ResizeMainControl(V2i(w,h),id)
+                        | _ -> Nop 
+                )
+                //onResize  (fun s -> OnResize(s,id))
             ] 
         ]            
 
@@ -1405,12 +1425,16 @@ module ViewerApp =
             | _ -> false
         ) m.ctrlFlag m.interaction
 
-    let viewInstrumentView (m: AdaptiveModel) = 
+    let viewInstrumentView (runtime : IRuntime) (id : string) (m: AdaptiveModel) = 
+        let frustum = m.frustum 
         let annotationsI, discsI = 
             DrawingApp.view 
                 m.scene.config 
                 mdrawingConfig 
                 (m.scene.viewPlans.instrumentCam)
+                frustum
+                runtime
+                (m.viewPortSize |> AMap.tryFind id |> AVal.map (Option.defaultValue V2i.II))
                 (allowAnnotationPicking m)
                 m.drawing
 
@@ -1435,12 +1459,19 @@ module ViewerApp =
             AVal.map2 Camera.create (m.scene.viewPlans.instrumentCam) m.scene.viewPlans.instrumentFrustum
         DomNode.RenderControl((instrumentControlAttributes m), icam, icmds, None) //AttributeMap.Empty
 
-    let viewRenderView (m: AdaptiveModel) = 
+    let viewRenderView (runtime : IRuntime) (id : string) (m: AdaptiveModel) = 
+
+        let frustum = AVal.map2 (fun o f -> o |> Option.defaultValue f) m.overlayFrustum m.frustum // use overlay frustum if Some()
+        let cam     = AVal.map2 Camera.create m.navigation.camera.view frustum
+
         let annotations, discs = 
             DrawingApp.view 
                 m.scene.config 
                 mdrawingConfig 
                 m.navigation.camera.view 
+                frustum
+                runtime
+                (m.viewPortSize |> AMap.tryFind id |> AVal.map (Option.defaultValue V2i.II))
                 (allowAnnotationPicking m)                 
                 m.drawing
             
@@ -1513,9 +1544,6 @@ module ViewerApp =
 
             let solText = 
                 MinervaApp.getSolBillboards m.minervaModel m.navigation.camera.view near |> Sg.map MinervaActions
-
-            let viewportFilteredText = 
-                MinervaApp.viewPortLabels m.minervaModel m.navigation.camera.view (ViewerUtils.frustum m) |> Sg.map MinervaActions
                 
             //let correlationLogs, _ =
             //    PRo3D.Correlations.CorrelationPanelsApp.viewWorkingLog 
@@ -1573,25 +1601,28 @@ module ViewerApp =
         let depthTested = 
             [linkingSg; annotationSg; minervaSg; heightValidationDiscs] |> Sg.ofList
 
-        let cmds    = ViewerUtils.renderCommands m.scene.surfacesModel.sgGrouped overlayed depthTested m
-        let frustum = AVal.map2 (fun o f -> o |> Option.defaultValue f) m.overlayFrustum m.frustum // use overlay frustum if Some()
-        let cam     = AVal.map2 Camera.create m.navigation.camera.view frustum
-        DomNode.RenderControl((renderControlAttributes m), cam, cmds, None)
+        let cmds  = ViewerUtils.renderCommands m.scene.surfacesModel.sgGrouped overlayed depthTested m
+        onBoot "attachResize('__ID__')" (
+            DomNode.RenderControl((renderControlAttributes id m), cam, cmds, None)
+        )
         
-    let view (m: AdaptiveModel) = //(localhost: string)
-       
+    let view (runtime : IRuntime) (m: AdaptiveModel) = //(localhost: string)
+
         let viewerDependencies = [
             { kind = Stylesheet;  name = "semui";           url = "https://cdn.jsdelivr.net/semantic-ui/2.2.6/semantic.min.css" }
             { kind = Stylesheet;  name = "semui-overrides"; url = "semui-overrides.css" }
             { kind = Script;      name = "semui";           url = "https://cdn.jsdelivr.net/semantic-ui/2.2.6/semantic.min.js" }
             { kind = Script;      name = "errorReporting";  url = "./errorReporting.js"  }
+
+            { kind = Script;      name = "resize";  url = "./ResizeSensor.js"  }
+            { kind = Script;      name = "resizeElem";  url = "./ElementQueries.js"  }
         ]
         
         let bodyAttributes = [style "background: #1B1C1E; height:100%; overflow-y:scroll; overflow-x:hidden;"] //overflow-y : visible
 
         page (
             fun request -> 
-                Gui.Pages.pageRouting viewerDependencies bodyAttributes m viewInstrumentView viewRenderView request
+                Gui.Pages.pageRouting viewerDependencies bodyAttributes m viewInstrumentView viewRenderView runtime request
         )
                    
     let threadPool (m: Model) =
@@ -1643,7 +1674,7 @@ module ViewerApp =
         App.start {
             unpersist = Unpersist.instance
             threads   = threadPool
-            view      = view //localhost
+            view      = view runtime //localhost
             update    = update runtime signature sendQueue messagingMailbox
             initial   = m
         }